backend/pyproject.toml
[project]
name = "make-it-lively-backend"
version = "0.1.0"
description = "FastAPI backend for Make It Lively"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "python-multipart>=0.0.9",
    "pillow>=10.4",
    "anthropic>=0.96.0",
    "replicate>=1.0",
    # Loads secrets from a local .env file at startup so developers do not
    # have to `export` ANTHROPIC_API_KEY / REPLICATE_API_TOKEN in every shell.
    "python-dotenv>=1.0",
    # Precise segmentation: contour extraction, GrabCut fallback, mask ops.
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["app", "tests"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "cv2.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
